service_templates/users/requirements.txt
aiotask-context==0.6.1
alembic==1.4.2
asn1crypto==1.4.0
astroid==2.4.2
attrs==19.3.0
azure-common==1.1.25
azure-core==1.7.0
azure-storage-blob==12.3.2
beautifulsoup4==4.9.1
boto3==1.14.38
botocore==1.17.38
bs4==0.0.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
cryptography==2.9.2
docutils==0.15.2
idna==2.9
inflect==4.1.0
iniconfig==1.0.1
isodate==0.6.0
isort==4.3.21
jmespath==0.10.0
lazy-object-proxy==1.4.3
logfmt==0.4
Mako==1.1.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.4.0
msrest==0.6.18
mypy==0.782
mypy-extensions==0.4.3
oauthlib==3.1.0
oscrypto==1.2.1
packaging==20.4
pluggy==0.13.1
psycopg2-binary==2.8.5
py==1.9.0
pycparser==2.20
pycryptodomex==3.9.8
PyJWT==1.7.1
pyjwt-rsa==0.3.9
pylint==2.5.3
pyOpenSSL==19.1.0
pyparsing==2.4.7
pytest==6.0.1
python-dateutil==2.8.1
python-editor==1.0.4
python-slugify==4.0.1
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.15.0
snowflake==0.0.3
snowflake-connector-python==2.2.10
soupsieve==2.0.1
sqlacodegen==2.3.0
SQLAlchemy==1.3.18
text-unidecode==1.3
toml==0.10.1
tornado==6.0.4
typed-ast==1.4.1
typing-extensions==3.7.4.2
urllib3==1.25.10
wrapt==1.12.1
